requirements.txt
asgiref==3.2.10
Django==3.1
django-filter==2.3.0
django-nested-admin==3.3.2
Pillow==7.2.0
pkg-resources==0.0.0
python-monkey-business==1.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
